+++ src/index.ts
@@ -58,6 +58,7 @@ export {
 	type AnySourceType,
 	type SourceTypeDefinition,
 } from "./tally/state/source/SourceType.js";
+export type { SourceBuilder } from "./tally/state/source/SourceBuilder.js";
 
 // Descriptor
 export type { AnyDescriptor, Descriptor } from "./tally/state/descriptor/Descriptor.js";
@@ -73,6 +74,7 @@ export {
 	type AnyDescriptorType,
 	type DescriptorTypeDefinition,
 } from "./tally/state/descriptor/DescriptorType.js";
+export type { DescriptorBuilder } from "./tally/state/descriptor/DescriptorBuilder.js";
 
 // Replication
 export type { ReplicationDefinition } from "./tally/replication/ReplicationDefinition.js";